pyproject.toml
[tool.poetry]
name = "calendar-ipynb"
version = "0.1.0"
description = ""
authors = ["Fahim Ali Zain <fahimalizain@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.161.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
ipywidgets = "^8.1.5"
jupyter = "^1.1.1"
notebook = "^7.3.2"
ipykernel = "^6.29.5"
matplotlib = "^3.10.0"
ipympl = "^0.9.6"
flake8 = "^7.1.2"
black = {extras = ["jupyter"], version = "^25.1.0"}
pandas = "^2.2.3"
mplcursors = "^0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88